(PC-23190)[API] fix: extract offerer_venue test
