+++ api/tests/routes/public/individual_offers/v1/get_booking_test.py
@@ -10,7 +10,7 @@ from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offers import factories as offers_factories
 from pcapi.utils import date as date_utils
 
-from .endpoints_test import create_offerer_provider_linked_to_venue
+from . import utils
 
 
 pytestmark = pytest.mark.usefixtures("db_session")
@@ -18,7 +18,7 @@ pytestmark = pytest.mark.usefixtures("db_session")
 
 class GetBookingByTokenReturns200Test:
     def test_key_has_rights_and_regular_product_offer(self, client):
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
         product_offer = offers_factories.ThingOfferFactory(
             venue=venue,
             description="Un livre de contrepèterie",
@@ -64,7 +64,7 @@ class GetBookingByTokenReturns200Test:
         }
 
     def test_key_has_rights_and_regular_event_offer(self, client):
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
         event_offer = offers_factories.EventOfferFactory(
             venue=venue,
             description="Un livre de contrepèterie",
@@ -122,7 +122,7 @@ class GetBookingByTokenReturns401Test:
 
 class GetBookingByTokenReturns403Test:
     def test_when_booking_not_confirmed(self, client):
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
         next_week = datetime.datetime.utcnow() + datetime.timedelta(weeks=1)
 
         offer = offers_factories.ThingOfferFactory(
@@ -149,7 +149,7 @@ class GetBookingByTokenReturns403Test:
 
     def test_when_booking_is_refunded(self, client):
         # Given
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
 
         offer = offers_factories.ThingOfferFactory(
             venue=venue,
@@ -173,7 +173,7 @@ class GetBookingByTokenReturns404Test:
         assert response.status_code == 404
 
     def test_key_has_no_rights_and_regular_offer(self, client):
-        create_offerer_provider_linked_to_venue()
+        utils.create_offerer_provider_linked_to_venue()
         venue = offerers_factories.VenueFactory()
         product_offer = offers_factories.ThingOfferFactory(
             venue=venue,
@@ -200,7 +200,7 @@ class GetBookingByTokenReturns404Test:
 
 class GetBookingByTokenReturns410Test:
     def test_when_booking_is_already_validated(self, client):
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
         product_offer = offers_factories.ThingOfferFactory(
             venue=venue,
         )
@@ -215,7 +215,7 @@ class GetBookingByTokenReturns410Test:
         assert response.json == {"booking": "This booking has already been validated"}
 
     def test_when_booking_is_cancelled(self, client):
-        venue, _ = create_offerer_provider_linked_to_venue()
+        venue, _ = utils.create_offerer_provider_linked_to_venue()
         product_offer = offers_factories.ThingOfferFactory(
             venue=venue,
         )
